Updates for amazon_ams as well as updating HG/QD coding in both the amazon customer and amazon po processes
