+++ amazon_ams/README.md
@@ -1,0 +1,7 @@
+How to updated:
+
+Make sure the new files saved to this folder: G:\SALES\Amazon\AMAZON ADVERTISING\MONTHLY REPORTS\MONTHLY REPORTS - PERFORMANCE BY ASIN\
+
+are saved in the same format, with the same columns.
+
+You'll need to updated to things.
