+++ amazon_customer_orders/loader/loader_item.py
@@ -9,9 +9,9 @@ def item_sql():
         i.ISBN
         ,i.SHORT_TITLE title
         ,CASE
-                WHEN i.PUBLISHING_GROUP IN('QDP-HGB','HGP-HGNA') THEN 'Hardie Grant Publishing'
-                WHEN i.PUBLISHING_GROUP IN('QDP-BOOK','QDP-GIFT','QDP-HBUK') THEN 'Quadrille'
-                ELSE i.PUBLISHER_CODE
+            WHEN left(i.PUBLISHING_GROUP,2) = 'HG' THEN 'Hardie Grant'
+            WHEN left(i.PUBLISHING_GROUP,3) = 'QDP' THEN 'Quadrille'
+            ELSE i.PUBLISHER_CODE
 	    END publisher
         ,case                             
             when i.PUBLISHER_CODE = 'Tourbillon' then 'TW'                      
@@ -28,8 +28,8 @@ def item_sql():
             when i.PUBLISHING_GROUP in('LAU-BIS') then 'LKBS'                          
             when i.PUBLISHER_CODE = 'Laurence King' and i.PRODUCT_TYPE = 'FT' then 'LKGI'                      
             when i.PUBLISHER_CODE = 'Laurence King' and i.PRODUCT_TYPE <> 'FT' then 'LKBK'         
-            when i.PUBLISHING_GROUP IN('QDP-HGB','HGP-HGNA') then 'HG' 
-            when i.PUBLISHING_GROUP IN('QDP-BOOK','QDP-GIFT','QDP-HBUK') then 'QD'
+            when left(i.PUBLISHING_GROUP,2) = 'HG' then 'HG' 
+            when left(i.PUBLISHING_GROUP,3) = 'QDP' then 'QD'
             when i.PUBLISHING_GROUP in('BAR-ART','BAR-ENT','BAR-LIF') then 'BAR'                  
             else i.PUBLISHING_GROUP                 
         end pgrp
